Added file upload and deletetion on edit/delete

+++ app/Http/Controllers/PostsController.php
@@ -3,10 +3,25 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Storage;
 use App\Post;
+
 
 class PostsController extends Controller
 {
+     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        $this->middleware('auth')->except(['index', 'show']);
+    }
+
+
+
+
     /**
      * Display a listing of the resource.
      *
@@ -38,16 +53,32 @@ class PostsController extends Controller
      */
     public function store(Request $request)
     {
-
-
         $validatedData = $request->validate([
-            'title' => 'required|unique:posts|max:255',
-            'article_body' => 'required',
+            'title'         => 'required|unique:posts|max:255',
+            'article_body'  => 'required',
+            'cover_image'   => 'image|nullable|max:1999'
         ]);
+        // Handle File Upload
+        if($request->hasFile('cover_image')) {
+            // Get Filename with extention 
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get Just Filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // GET EXTENSION
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            // Link store with public by artisan 
+
+        } else {
+            $fileNameToStore = "noimage.jpg";
+        }
         $post = new Post;
         $post->title = $request->input('title'); 
         $post->body = $request->input('article_body'); 
         $post->user_id = auth()->user()->id;
+        $post->cover_image = $fileNameToStore;
         $post->save();
         
         return redirect()->to('posts')->with('success', 'Post created');
@@ -74,6 +105,10 @@ class PostsController extends Controller
     public function edit($id)
     {
         $post = Post::find($id);
+        // Check correct user
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
         return view('posts.edit')->with('post', $post);
     }
 
@@ -89,10 +124,28 @@ class PostsController extends Controller
         $validatedData = $request->validate([
             'title' => 'required|max:255',
             'article_body' => 'required',
+            'cover_image'   => 'image|nullable|max:1999'
         ]);
+
+        if($request->hasFile('cover_image')) {
+            // Get Filename with extention 
+            $filenameWithExt = $request->file('cover_image')->getClientOriginalName();
+            // Get Just Filename
+            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+            // GET EXTENSION
+            $extension = $request->file('cover_image')->getClientOriginalExtension();
+            $fileNameToStore = $filename.'_'.time().'.'.$extension;
+            // Upload Image
+            $path = $request->file('cover_image')->storeAs('public/cover_images', $fileNameToStore);
+            // Link store with public by artisan 
+        } 
         $post = Post::find($id);
         $post->title = $request->input('title'); 
         $post->body = $request->input('article_body'); 
+        if($request->hasFile('cover_image')) {
+            Storage::delete('public/cover_images/'.$post->cover_image);
+            $post->cover_image = $fileNameToStore;
+        }
         $post->save();
         
         return redirect()->to('posts')->with('success', 'Post Updated');
@@ -108,6 +161,13 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('posts')->with('error', 'Unauthorized Page');
+        }
+        if($post->cover_image !== 'noimage.jpg'){
+            // Delete Image
+            Storage::delete('public/cover_images/'.$post->cover_image);
+        }
         $post->delete();
         return redirect()->to('posts')->with('success', 'Post Deleted');
